+++ mayan/apps/documents/views/document_views.py
@@ -596,9 +596,21 @@ class DocumentUpdatePageCountView(MultipleObjectConfirmActionView):
         return result
 
     def object_action(self, form, instance):
-        task_update_page_count.apply_async(
-            kwargs={'version_id': instance.latest_version.pk}
-        )
+        latest_version = instance.latest_version
+        if latest_version:
+            task_update_page_count.apply_async(
+                kwargs={'version_id': latest_version.pk}
+            )
+        else:
+            messages.error(
+                self.request, _(
+                    'Document "%(document)s" is empty. Upload at least one '
+                    'document version before attempting to detect the '
+                    'page count.'
+                ) % {
+                    'document': instance,
+                }
+            )
 
 
 class DocumentTransformationsClearView(MultipleObjectConfirmActionView):